pyproject.toml
[tool.poetry]
name = "insta-stamp-flask"
version = "0.1.0"
description = ""
authors = ["조현일 <himzei@gmail.com>"]
readme = "README.md"
packages = [{include = "insta_stamp_flask"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.7"
djangorestframework = "^3.14.0"
beautifulsoup4 = "^4.11.2"
requests = "^2.28.2"
selenium = "^4.8.2"
webdriver-manager = "^3.8.5"
lxml = "^4.9.2"
Pillow = "^9.4.0"
insta-scrape = "^2.1.2"
django-cors-headers = "^3.14.0"
gunicorn = "^20.1.0"
whitenoise = "^6.4.0"
django-environ = "^0.10.0"
django-pymssql = "^1.7.1"
mssql-django = "^1.2"
pyodbc = "^4.0.35"
django-mssql-backend = "^2.8.1"
pymssql = "^2.2.7"
django-odbc = "^0.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
